requirements.txt
aioredis==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
channels==3.0.4
channels-redis==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coverage==7.4.3
cryptography==42.0.4
daphne==3.0.2
Django==4.0
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
ecdsa==0.18.0
exceptiongroup==1.2.0
hiredis==2.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
kombu==5.3.5
msgpack==1.0.7
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
pytest==8.1.0
python-dateutil==2.9.0
python-jose==3.3.0
pytz==2024.1
redis==5.0.1
rsa==4.9
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
